+++ L2J_Mobius_CT_2.6_HighFive/java/com/l2jmobius/gameserver/model/actor/L2Attackable.java
@@ -363,7 +363,7 @@ public class L2Attackable extends L2Npc
 			final Map<L2PcInstance, DamageDoneInfo> rewards = new ConcurrentHashMap<>();
 			
 			L2PcInstance maxDealer = null;
-			int maxDamage = 0;
+			long maxDamage = 0;
 			long totalDamage = 0;
 			// While Iterating over This Map Removing Object is Not Allowed
 			// Go through the _aggroList of the L2Attackable
@@ -379,7 +379,7 @@ public class L2Attackable extends L2Npc
 				if (attacker != null)
 				{
 					// Get damages done by this attacker
-					final int damage = info.getDamage();
+					final long damage = info.getDamage();
 					
 					// Prevent unwanted behavior
 					if (damage > 1)
@@ -429,7 +429,7 @@ public class L2Attackable extends L2Npc
 					final L2PcInstance attacker = reward.getAttacker();
 					
 					// Total amount of damage done
-					final int damage = reward.getDamage();
+					final long damage = reward.getDamage();
 					
 					// Get party
 					final L2Party attackerParty = attacker.getParty();
@@ -485,7 +485,7 @@ public class L2Attackable extends L2Npc
 					else
 					{
 						// share with party members
-						int partyDmg = 0;
+						long partyDmg = 0;
 						double partyMul = 1;
 						int partyLvl = 0;
 						
@@ -1328,7 +1328,7 @@ public class L2Attackable extends L2Npc
 	 * @param totalDamage The total damage done
 	 * @return
 	 */
-	private double[] calculateExpAndSp(int charLevel, int damage, long totalDamage)
+	private double[] calculateExpAndSp(int charLevel, long damage, long totalDamage)
 	{
 		final int levelDiff = charLevel - getLevel();
 		double xp = 0;
@@ -1672,7 +1672,7 @@ public class L2Attackable extends L2Npc
 	/*
 	 * Return vitality points decrease (if positive) or increase (if negative) based on damage. Maximum for damage = maxHp.
 	 */
-	public float getVitalityPoints(int damage)
+	public float getVitalityPoints(long damage)
 	{
 		// sanity check
 		if (damage <= 0)
