+++ L2J_Mobius_CT_2.6_HighFive/java/com/l2jmobius/gameserver/model/actor/instance/L2GrandBossInstance.java
@@ -105,7 +105,7 @@ public final class L2GrandBossInstance extends L2MonsterInstance
 	}
 	
 	@Override
-	public float getVitalityPoints(int damage)
+	public float getVitalityPoints(long damage)
 	{
 		return -super.getVitalityPoints(damage) / 100;
 	}
