+++ L2J_Mobius_Classic_2.1_Zaken/java/com/l2jmobius/gameserver/model/actor/L2Attackable.java
@@ -348,7 +348,7 @@ public class L2Attackable extends L2Npc
 			final Map<L2PcInstance, DamageDoneInfo> rewards = new ConcurrentHashMap<>();
 			
 			L2PcInstance maxDealer = null;
-			int maxDamage = 0;
+			long maxDamage = 0;
 			long totalDamage = 0;
 			// While Iterating over This Map Removing Object is Not Allowed
 			// Go through the _aggroList of the L2Attackable
@@ -359,7 +359,7 @@ public class L2Attackable extends L2Npc
 				if (attacker != null)
 				{
 					// Get damages done by this attacker
-					final int damage = info.getDamage();
+					final long damage = info.getDamage();
 					
 					// Prevent unwanted behavior
 					if (damage > 1)
@@ -450,7 +450,7 @@ public class L2Attackable extends L2Npc
 					final L2PcInstance attacker = reward.getAttacker();
 					
 					// Total amount of damage done
-					final int damage = reward.getDamage();
+					final long damage = reward.getDamage();
 					
 					// Get party
 					final L2Party attackerParty = attacker.getParty();
@@ -524,7 +524,7 @@ public class L2Attackable extends L2Npc
 					else
 					{
 						// share with party members
-						int partyDmg = 0;
+						long partyDmg = 0;
 						double partyMul = 1;
 						int partyLvl = 0;
 						
@@ -1307,7 +1307,7 @@ public class L2Attackable extends L2Npc
 	 * @param totalDamage The total damage done
 	 * @return
 	 */
-	private double[] calculateExpAndSp(int charLevel, int damage, long totalDamage)
+	private double[] calculateExpAndSp(int charLevel, long damage, long totalDamage)
 	{
 		final int levelDiff = charLevel - getLevel();
 		double xp = 0;
